Migrate layout and index page to the app router
Add app directory and define layout and page

Move index page to app/home-page and remove getStaticProps

Use Edge runtime to fix build errors when backend is not connected

Add api/posts/edge to getPostsLocal using fetch
(axios doesn't work with edge runtime, https://github.com/axios/axios/issues/5523)

Add `use client` to all client-side components
