+++ frontend/src/components/posts/list.tsx
@@ -1,3 +1,5 @@
+"use client"
+
 import Link from "next/link"
 import { FC } from "react"
