+++ frontend/tsconfig.json
@@ -16,8 +16,13 @@
     "incremental": true,
     "paths": {
       "@/*": ["./src/*"]
-    }
+    },
+    "plugins": [
+      {
+        "name": "next"
+      }
+    ]
   },
-  "include": ["./next-env.d.ts", "**/*.ts", "**/*.tsx", "src/components/categoryMultiselect.js"],
+  "include": ["./next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts"],
   "exclude": ["node_modules"]
 }
